python/requirements.txt
oauthlib==0.6.1
py==1.4.20
pytest==2.5.2
python-dateutil==2.2
requests==2.2.1
requests-oauthlib==0.4.0
